Core functionality added
Functionality is in place with Recording scene. UI needs to be fixed for the non-saving portion. Playback scene is added and works with limited capabilities. Goal is to customize how many AudioStreamPlayers the user wants to have playing at once, and also things like specific volume for each source and etc.